Fix price column type in stocks table from INT to NUMERIC(10, 2)

+++ docker/postgres/init.sql
@@ -1,7 +1,7 @@
 CREATE TABLE IF NOT EXISTS stocks (
     id SERIAL PRIMARY KEY,
     name TEXT NOT NULL,
-    price INT NOT NULL,
+    price NUMERIC(10, 2) NOT NULL,
     company TEXT NOT NULL,
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
